cleans up app structure and fixes a few issues and adds centering to ChooseProgram component
