+++ src/components/buttons/ButtonPrimary.jsx
@@ -1,0 +1,17 @@
+// Buttons mit den drei Gradients
+
+export default function ButtonPrimary({ children, className, bg = "grad1" }) {
+  const bgStyles = {
+    grad1: "from-g1a to-g1b",
+    grad2: " from-g2a to-g2b",
+    grad3: "from-g3d from-0% via-g3c via-[21.35%] to-g3b to-[92.71%]",
+  };
+
+  return (
+    <button
+      className={`${className} h-11 w-56 rounded-3xl bg-gradient-to-br ${bgStyles[bg]}`}
+    >
+      {children}
+    </button>
+  );
+}
