--- src/pages/error-page.jsx
@@ -1,14 +1,0 @@
-import { useRouteError } from "react-router-dom";
-
-export default function ErrorPage() {
-    const error = useRouteError();
-    console.error(error);
-
-    return (
-        <div id="error-page">
-            <h1>Sorry!</h1>
-            <p>An unexpected error has occured.</p>
-            <i>{error.statusText || error.message}</i>
-        </div>
-    );
-}
