+++ src/pages/home.jsx
@@ -1,6 +1,7 @@
 import DefaultLayout from "../layouts/DefaultLayout.jsx"; //explizit die Komponente importieren, nicht die ganze Datei; NavBar Variable wird heir deklariert
 import ChooseProgram from "../components/choose-programm.jsx";
 import ShowOverview from "../components/overview-workout.jsx";
+import ButtonGrad1, { ButtonGrad2, ButtonGrad3 } from "../components/buttons/button-primary.jsx";
 
 const user = {
   name: "Maxine Musterfrau",
@@ -15,7 +16,7 @@ function Home() {
           Hi {user.name}!
         </h1>
       </div>
-      <div className="mb-10 flex justify-center">
+      <div className="mb-10 justify-center flex">
         <img src="../startbild.png"></img>
       </div>
       {/* Bedingung true? dann ShowOverview, else ChooseProgram */}
